BLTPrivateMacros: match language suffixes by regex
This commit changes the language source code extension detection code
to match extensions via regular expressions on filenames instead of
matching by using the output of `get_filename_component(... ... EXT)`
and checking to see if that output is in a list of filename
extensions.

`get_filename_component(... ... EXT)` will output the string formed
from the leftmost period in the filename (ignoring directory
prefixes), plus all characters to the right of that period. A
challenge with this approach occurs when files are named, e.g.,
`quadrature_rule.order2.c` because
`get_filename_component(... ... EXT)` applied to those files will
return, e.g., `.order2.c`, which will not match `.c`. This limitation
forces users to add extensions to BLT variables like
`BLT_C_FILE_EXTS`. While this approach is tractable if a small number
of extensions must be added, for larger numbers of "custom"
extensions, this approach becomes cumbersome for users.

The approach in this commit instead forms a regex of the form, e.g.,
for C-like extensions,

```
^.*(\.cpp|\.hpp|\.cxx|\.cxx|\.hxx|\.c|\.h|\.cc|\.hh|\.inl|\.cu|\.cuh)$
```

by looping over file extension settings, contatenating the list with
pipes, and escaping the period in file extensions appropriately. This
regex will match the `.c` part of `quadrature_rule.order2.c` and
appropriately identify it as a C-like source file.